requirements.txt
joblib==1.0.1
pylab_sdk==1.3.2
scipy==1.6.3
pandas==1.2.0
numpy==1.20.2
Flask==1.1.2
scikit_learn==0.24.2
gunicorn
requests